Update mainnet deployment (#7865)
Fix link to profiling gas section to reflect late renaming (#7525)
Fix e2e tests following change to --skip-faucet (#7561)
fix ghost update (#7491)
Fix unnecessary metric dimension (#7388)
fix yaml quotation (#7620)
Fix legacy testing for gas (#7641)
Fix all simple maps to be vector rather than Btreemap (#7642)
Fix txn emitter client filtering (#7433)
fix internal (#7673)
Fix links to migrated Rust resources from Rust Coding Guidelines (#7688)
fix tip  format (#7668)
Fix links to TypeScript files reorganized in PR #6993 (#7690)
fix (#7643)
Fix the CFG bug from Zelig (#73)
Fix fmt (#74)
Fix some typos in staking docs (#7743)
Fix lint.py shell invocation (#7751)
Fix grammar / typos in staking pool operations doc (#7744)
fix diem specs (#7754)
Fix logging init in lint.py and find_latest_image.py (#7755)
Fix Poetry dependency on Python and update lockfile (#7775)
Fix three region simulation (#7757)
Fix cached-packages failure. (#7811)
fix the abstract interpreter bypass bug (#7810)
fix (#7818)
Fix MultisigTransactionPayload deserialization (#7836)
Fixes for PFNs (#7833)

+++ crates/aptos-telemetry/src/utils.rs
@@ -1,4 +1,4 @@
-// Copyright (c) Aptos
+// Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
 #![forbid(unsafe_code)]
@@ -61,7 +61,7 @@ pub fn sum_all_histogram_sums(metric_families: &Vec<MetricFamily>) -> f64 {
     let mut count_sum = 0.0;
     for metric_family in metric_families {
         for metric in metric_family.get_metric() {
-            count_sum += metric.get_histogram().get_sample_sum() as f64
+            count_sum += metric.get_histogram().get_sample_sum()
         }
     }
     count_sum
