+++ types/src/proof/position/mod.rs
@@ -1,4 +1,5 @@
-// Copyright (c) Aptos
+// Copyright © Aptos Foundation
+// Parts of the project are originally copyright © Meta Platforms, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
 //! This module provides an abstraction for positioning a node in a binary tree,
@@ -166,12 +167,12 @@ impl Position {
 
     pub fn root_from_leaf_count(leaf_count: LeafCount) -> Self {
         assert!(leaf_count > 0);
-        Self::root_from_leaf_index((leaf_count - 1) as u64)
+        Self::root_from_leaf_index(leaf_count - 1)
     }
 
     pub fn root_level_from_leaf_count(leaf_count: LeafCount) -> u32 {
         assert!(leaf_count > 0);
-        let index = (leaf_count - 1) as u64;
+        let index = leaf_count - 1;
         MAX_ACCUMULATOR_PROOF_DEPTH as u32 + 1 - index.leading_zeros()
     }
 
@@ -456,7 +457,7 @@ impl Iterator for FrozenSubtreeSiblingIterator {
         self.remaining_new_leaves -= next_subtree_leaves;
 
         Some(Position::from_inorder_index(
-            (first_leaf_index + last_leaf_index) as u64,
+            first_leaf_index + last_leaf_index,
         ))
     }
 }
